create score server top widget
